pyproject.toml
[project]
name = "intership-assignment"
version = "0.1.0"
description = "Educational Content Generator with AI Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.0",
    "groq>=0.4.0",
    "python-dotenv>=1.0.0"
]
